requirements-dev.txt
# Pragma Trading Bot - Development & Testing Requirements
# Install with: pip install -r requirements-dev.txt
# This includes all production requirements plus dev tools

# Include production requirements
-r requirements.txt

# Testing Framework
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-asyncio==0.25.2
pytest-timeout==2.3.1
hypothesis==6.140.3

# Code Quality & Linting
black==25.9.0
flake8==7.3.0
mypy==1.13.0
pylint==3.3.1
isort==5.13.2
bandit==1.8.0  # Security linter

# Type Stubs
types-requests==2.32.0.20241016
pandas-stubs==2.2.2.240909

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.2
myst-parser==4.0.0

# Development Tools
ipython==8.30.0
jupyter==1.1.1
notebook==7.3.2

# Visualization (for analysis)
matplotlib==3.10.0
plotly==5.24.1
seaborn==0.13.2

# Profiling & Debugging
memory-profiler==0.61.0
line-profiler==4.2.0
py-spy==0.4.0

# Pre-commit hooks (optional)
pre-commit==4.1.0

# Coverage reporting
coverage[toml]==7.6.9
